+++ CHANGELOG.md
@@ -8,6 +8,9 @@
     - Add new option `htmlEnabled` for HTML message support (issue #51)
         - `true`: render message with `JLabel`
         - `false`: render message with `JTextArea` (default)
+    - When showing a toast, it now returns a unique toast ID. (issue #52) (PR #53)
+        - Format: `toast-{toast_count}`
+        - Example: `toast-1`, `toast-2`
 
 ## [2.6.0] - 2026-02-19